Keep Today card tests within CI census
